Add maven-pmd-plugin
PMD is a configurable tool for static code analysis. In the initial
configuration is detects exception-related issues and the massive
duplication of string literals. It's executed in the validate phase of
the build.

NullPointerExceptions are used where IllegalArgumentException should be
used (unfortunately this pattern is encouraged by the JSF specification,
e.g. in FacesValidator), but since the careful removal of such flaws
takes time they're initially marked to pass the validation, but can be
found easily by searching for the SupporessWarnings annotation.

Throwing of Exception has been removed where possible since it has no
advantage over throwing the specific Exception subclass and catching
Exception is almost never right. The catching has been marked to pass
validation like NullPointerExceptions.

String literals shouldn't be repeated excessively because that leads to
the risk of typos (even if it's very small) and slows down compilation.